A small change in the writeFile functions fixed the corrupted compression issue on windows. I needed to set ios_base::binary on the ofstreams. My serialization classes already did this, but my new write functions did not.
git-svn-id: file:///mnt/svn/rtmath@1088 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/src/ddOutput.cpp
@@ -124,12 +124,12 @@ namespace rtmath {
 				{
 					if (!thetas.count(sca->theta())) thetas.insert(sca->theta());
 				}
-				numThetas = thetas.size();
+				numThetas = (float) thetas.size();
 			}
 			/// \todo Put this elsewhere, for the ddOutPutEnsembleDDSCAT function to also use
 			for (auto &sca : res->scas)
 			{
-				double weight = 0;
+				float weight = 0;
 				auto fApprox = [](double a, double b) -> bool
 				{
 					if (abs((a-b)/a) < 0.001) return true;
